feat: 코드리팩토링, 구조개선
