+++ .dockerignore
@@ -1,5 +1,9 @@
 node_modules/
 .git/
+.gitignore
+*.swp
+.DS_Store
+.idea/
 .vscode/
 *.log
 *.json
@@ -8,3 +12,4 @@ temp/
 data/
 __pycache__/
 README.md
+!pyproject.toml/requirements.txt
