+++ .gitignore
@@ -208,6 +208,7 @@ __marimo__/
 
 # temp folder
 temp/
+dev/
 .vscode/
 
 # Data
