+++ service/maplestory_command.py
@@ -1,24 +1,21 @@
-"""
-
-디스코드 봇에서 Nexon Open API를 사용하여 메이플스토리 관련 명령어를 처리하는 모듈
-
-Reference: https://openapi.nexon.com/
-
-"""
-
 import discord
-import hashlib
 from discord.ext import commands
 
-from bs4 import BeautifulSoup
-from urllib.parse import quote
-import matplotlib.pyplot as plt
 import pandas as pd
 import numpy as np
+import io
+from matplotlib import pyplot as plt
+from bs4 import BeautifulSoup
 
-from service.common import log_command, parse_iso_string, preprocess_int_with_korean
-from service.api_utils import *
-from config import NEXON_API_HOME
+from service.maplestory_utils import *
+
+from bot_logger import log_command
+from utils.image import get_image_bytes
+from utils.text import preprocess_int_with_korean
+from utils.time import parse_iso_string
+
+from exceptions.api_exceptions import *
+
 
 @log_command
 async def api_basic_info(ctx: commands.Context, character_name: str) -> None:
@@ -204,6 +201,7 @@ async def api_basic_info(ctx: commands.Context, character_name: str) -> None:
     else:
         embed.colour = discord.Colour.from_rgb(128, 128, 128)
     await ctx.send(embed=embed)
+
 
 @log_command
 async def api_pcbang_notice(ctx: commands.Context) -> None:
@@ -432,6 +430,7 @@ async def api_sunday_notice(ctx: commands.Context) -> None:
     # 공지사항이 없을 때
     else:
         await ctx.send("썬데이 이벤트 공지사항이 아직 없어양!!\n매주 금요일 오전 10시에 업데이트 되니 참고해양!!")
+
 
 @log_command
 async def api_detail_info(ctx: commands.Context, character_name: str) -> None:
@@ -934,738 +933,6 @@ async def api_ability_info(ctx: commands.Context, character_name: str) -> None:
         embed.set_footer(text=f"어빌리티 최대값은 숫자 뒤 괄호안에 표시되어 있어양")
         await ctx.send(embed=embed)
 
-@log_command
-async def api_weather(ctx: commands.Context, location_name: str) -> None:
-    """현재 지역의 날씨 정보, 예보 정보를 가져오는 명령어
-
-    Args:
-        ctx (commands.Context): Discord 명령어 컨텍스트
-        location_name (str): 지역 이름/주소
-
-    Raises:
-        Exception : 지역정보 조회, 날씨 조회 실패 시 발생
-
-    Reference:
-        [지역 정보 조회 API (KAKAO developers)](https://developers.kakao.com/docs/latest/ko/local/dev-guide#search-by-address)
-        [날씨 조회 API (Data.go.kr)](https://www.data.go.kr/data/15084084/openapi.do)
-    """
-    if ctx.message.author.bot:
-        return
-    
-    try:
-        # 지역 정보 조회
-        location_data = get_local_info(local_name=location_name)
-        local_type = location_data.get('address_type')
-        if local_type == "REGION":
-            local_address_name = location_data.get('address_name')
-            local_x: str = location_data.get('x')
-            local_y: str = location_data.get('y')
-        else:
-            local_road_address: dict = location_data.get('road_address')
-            local_address_1 = local_road_address.get('region_1depth_name')
-            local_address_2 = local_road_address.get('region_2depth_name')
-            local_address_3 = local_road_address.get('region_3depth_name')
-            local_address_name = f"{local_address_1} {local_address_2} {local_address_3}"
-            local_x: str = local_road_address.get('x')
-            local_y: str = local_road_address.get('y')
-    except KKO_LOCAL_API_ERROR as e:
-        await ctx.send(f"해당 지역의 정보를 검색하는 중에 오류가 발생했어양!")
-        raise KakaoAPIError(str(e))
-    except KKO_NO_LOCAL_INFO as e:
-        await ctx.send(f"해당 지역의 정보를 찾을 수 없어양!")
-        raise KakaoAPIError(str(e))
-    
-    try:
-        # 날씨 정보 조회
-        weather_info = get_weather_info(local_x, local_y)
-    except WTH_API_INTERNAL_ERROR:
-        await ctx.send(f"날씨 정보를 가져오는 중에 오류가 발생했어양!")
-        raise WeatherAPIError("Internal server error")
-    except WTH_API_DATA_ERROR:
-        await ctx.send(f"날씨 API 데이터에 문제가 발생했어양!")
-        raise WeatherAPIError("Data error")
-    except WTH_API_DATA_NOT_FOUND:
-        await ctx.send(f"해당 지역의 날씨 정보를 찾을 수 없어양!")
-        raise WeatherAPIError("Data not found")
-    except WTH_API_HTTP_ERROR:
-        await ctx.send(f"날씨 API 요청 중에 오류가 발생했어양!")
-        raise WeatherAPIError("HTTP error")
-    except WTH_API_TIMEOUT:
-        await ctx.send(f"날씨 데이터 가져오는데 시간이 초과되었어양!")
-        raise WeatherAPIError("Timeout error")
-    except WTH_API_INVALID_PARAMS:
-        await ctx.send(f"날씨 API 요청 파라미터가 잘못되었어양!")
-        raise WeatherAPIError("Invalid params")
-    except WTH_API_INVALID_REGION:
-        await ctx.send(f"해당 지역은 날씨 API에서 지원하지 않아양!")
-        raise WeatherAPIError("Invalid region")
-    except WTH_API_DEPRECATED:
-        await ctx.send(f"더 이상 지원되지 않는 기능이에양!")
-        raise WeatherAPIError("Deprecated feature")
-    except WTH_API_UNAUTHORIZED:
-        await ctx.send(f"날씨 API 서비스 접근 권한이 없어양!")
-        raise WeatherAPIError("Unauthorized access to API")
-    except WTH_API_KEY_TEMP_ERROR:
-        await ctx.send(f"날씨 API 키가 임시로 제한되었어양!")
-        raise WeatherAPIError("Temporary API key restriction")
-    except WTH_API_KEY_LIMIT_EXCEEDED:
-        await ctx.send(f"날씨 API 키의 요청 한도를 초과했어양!")
-        raise WeatherAPIError("API key request limit exceeded")
-    except WTH_API_KEY_INVALID:
-        await ctx.send(f"날씨 API 키가 유효하지 않아양!")
-        raise WeatherAPIError("Invalid API key")
-    except WTH_API_KEY_EXPIRED:
-        await ctx.send(f"날씨 API 키가 만료되었어양!")
-        raise WeatherAPIError("Expired API key")
-    except WeatherAPIError:
-        await ctx.send(f"날씨 API 요청 중에 오류가 발생했어양!")
-        raise WeatherAPIError("Weather API error")
-    except Exception as e:
-        await ctx.send(f"날씨 정보를 가져오는 중에 알 수 없는 오류가 발생했어양!")
-        raise WeatherAPIError(str(e))
-
-    # 날씨 데이터 전처리 - 실황 정보
-    kst_now: datetime = datetime.now(tz=timezone("Asia/Seoul"))
-    ncst_info: dict = weather_info.get("ncst")
-    ncst_time: str = ncst_info.get("ncst_time", "몰라양")
-
-    # 현재 온도
-    val_temperature: str = ncst_info.get("temperature")
-    if "알수없음" in val_temperature:
-        current_temp: str = "몰라양"
-    else:
-        current_temp: str = f"{val_temperature.strip()}"
-    
-    # 현재 습도
-    val_humidity: str = ncst_info.get("humidity")
-    if "알수없음" in val_humidity:
-        current_humidity: str = "몰라양"
-    else:
-        current_humidity: str = f"{val_humidity.strip()}"
-
-    # 현재 풍속
-    val_wind_speed: str = ncst_info.get("wind_speed") # 0.0 m/s
-    if "알수없음" in val_wind_speed:
-        wind_speed_text: str = "몰라양"
-    else:
-        wind_speed_text: str = f"{val_wind_speed.strip()}"
-        val_wind_speed_float: float = float(val_wind_speed.replace("m/s", "").strip())
-        if val_wind_speed_float >= 4.0 and val_wind_speed_float < 9.0:
-            wind_speed_text: str = f"{val_wind_speed.strip()} (약간 강한 바람)"
-        elif val_wind_speed_float >= 9.0 and val_wind_speed_float < 14.0:
-            wind_speed_text: str = f"{val_wind_speed.strip()} (강한 바람)"
-        elif val_wind_speed_float >= 14.0 and val_wind_speed_float < 20.0:
-            wind_speed_text: str = f"{val_wind_speed.strip()} (매우 강한 바람)"
-        elif val_wind_speed_float >= 20.0:
-            wind_speed_text: str = f"{val_wind_speed.strip()} (폭풍 수준의 바람)"
-            
-    # 현재 풍향
-    val_wind_direction: str = ncst_info.get("wind_direction")
-    if "알수없음" in val_wind_direction:
-        current_wind_direction: str = "몰라양"
-    else:
-        current_wind_direction: str = f"{val_wind_direction.strip()}"
-
-    # 현재 강수 형태
-    val_rain_type: str = ncst_info.get("rainsnow_type")
-    if "알수없음" in val_rain_type:
-        current_rain_type: str = "몰라양"
-    else:
-        current_rain_type: str = f"{val_rain_type.strip()}"
-
-    # 현재 1시간 강수량 (비 또는 눈이 오는 경우 제공)
-    if val_rain_type in ["없음", "알수없음"]:
-        current_rain_flag: bool = False
-    else:
-        current_rain_flag: bool = True
-
-    if current_rain_flag:
-        val_rain_1h: str = ncst_info.get("rain_1h_value")
-        val_rain_1h_desc: str = ncst_info.get("rain_1h_desc")
-        val_rain_1h_float: float = float(val_rain_1h)
-        if val_rain_1h_float >= 30.0 and val_rain_1h_float < 50.0:
-            val_rain_1h_float_text = "들풍과 천둥, 번개를 동반한 비가 내릴 수 있어양."
-        elif val_rain_1h_float >= 50.0 and val_rain_1h_float < 70.0:
-            val_rain_1h_float_text = "도로가 침수될 수 있고, 차량 운행이 어려울 수 있어양."
-        elif val_rain_1h_float >= 70.0:
-            val_rain_1h_float_text = "심각한 피해가 발생할 수 있어양. 이불 밖은 위험해양!"
-        else:
-            val_rain_1h_float_text = "우산을 챙기세양. 비가 내릴 수 있어양."
-
-        current_rain_text: str = (
-            f"현재 1시간 강수량이 {val_rain_1h}mm 이에양.\n"
-            f"{val_rain_1h_float_text}"
-        )
-        current_rain_type: str = f"비 ({val_rain_1h_desc})"
-    else:
-        current_rain_text: str = ""
-    # 실황 정보 메세지 생성
-    ncst_hhmm: str = kst_now.strftime("%H:%M")
-    ncst_head: str = f"📍 현재 날씨 정보 ({ncst_hhmm})\n" # HH:MM
-    ncst_text: str = (
-        f"{ncst_head}"
-        f"**현재 기온**: {current_temp}\n"
-        f"**현재 습도**: {current_humidity}\n"
-        f"**현재 풍속**: {current_wind_direction}풍 {wind_speed_text}\n"
-        f"**강수 형태**: {current_rain_type}\n"
-    )
-
-    # 날씨 데이터 전처리 - 예보 정보
-    fcst_info: dict = weather_info.get("fcst")
-    fcst_time: str = fcst_info.get("fcst_time", "몰라양")
-
-    # N시간 후 예보 정보 설정
-    time_interval_hour_t1: int = 2
-    time_interval_hour_t2: int = 4
-
-    fcst_base_time: datetime = kst_now.replace(minute=0, second=0, microsecond=0)
-    after_t1_time: datetime = fcst_base_time + timedelta(hours=time_interval_hour_t1)
-    after_t1_time_str: str = after_t1_time.strftime("%Y%m%d-%H%M")
-    after_t2_time: datetime = fcst_base_time + timedelta(hours=time_interval_hour_t2)
-    after_t2_time_str: str = after_t2_time.strftime("%Y%m%d-%H%M")
-
-    # "SKY" : 하늘상태 (0~5: 맑음, 6~8: 구름많음, 9~10: 흐림)
-    fcst_sky: list[dict] = fcst_info.get("SKY", [])
-    if fcst_sky:
-        fcst_sky_text_t1: str = ""
-        fcst_sky_text_t2: str = ""
-        for sky in fcst_sky:
-            fcst_datetime_str: str = sky.get("fcst_datetime_str")
-            # t1/t2 시간 후 예보만 추출
-            if fcst_datetime_str == after_t1_time_str:
-                val_sky_t1: str = sky.get("value", "몰라양")
-                imo_sky_t1: str = get_sky_icon(val_sky_t1)
-                fcst_sky_text_t1: str = f"**하늘 상태**: {imo_sky_t1}\n"
-            elif fcst_datetime_str == after_t2_time_str:
-                val_sky_t2: str = sky.get("value", "몰라양")
-                imo_sky_t2: str = get_sky_icon(val_sky_t2)
-                fcst_sky_text_t2: str = f"**하늘 상태**: {imo_sky_t2}\n"
-    else:
-        fcst_sky_text_t1: str = ""
-        fcst_sky_text_t2: str = ""
-
-    # T1H : 기온 (단위: ℃)
-    fcst_t1h: list[dict] = fcst_info.get("T1H", [])
-    if fcst_t1h:
-        fcst_t1h_text_t1: str = ""
-        fcst_t1h_text_t2: str = ""
-        for t1h in fcst_t1h:
-            fcst_datetime_str: str = t1h.get("fcst_datetime_str")
-            # t1/t2 시간 후 예보만 추출
-            if fcst_datetime_str == after_t1_time_str:
-                val_t1h_t1: str = t1h.get("value", "몰라양")
-                fcst_t1h_text_t1 = f"**기온**: {val_t1h_t1}℃\n"
-            elif fcst_datetime_str == after_t2_time_str:
-                val_t1h_t2: str = t1h.get("value", "몰라양")
-                fcst_t1h_text_t2 = f"**기온**: {val_t1h_t2}℃\n"
-    else:
-        fcst_t1h_text_t1: str = ""
-        fcst_t1h_text_t2: str = ""
-
-    # REH : 습도 (단위: %)
-    fcst_reh: list[dict] = fcst_info.get("REH", [])
-    if fcst_reh:
-        fcst_reh_text_t1: str = ""
-        fcst_reh_text_t2: str = ""
-        for reh in fcst_reh:
-            fcst_datetime_str: str = reh.get("fcst_datetime_str")
-            # t1/t2 시간 후 예보만 추출
-            if fcst_datetime_str == after_t1_time_str:
-                val_reh_t1: str = reh.get("value", "몰라양")
-                fcst_reh_text_t1 = f"**습도**: {val_reh_t1}%\n"
-            elif fcst_datetime_str == after_t2_time_str:
-                val_reh_t2: str = reh.get("value", "몰라양")
-                fcst_reh_text_t2 = f"**습도**: {val_reh_t2}%\n"
-    else:
-        fcst_reh_text_t1: str = ""
-        fcst_reh_text_t2: str = ""
-
-    # VEC / WSD : 풍향 / 풍속
-    fcst_vec: list[dict] = fcst_info.get("VEC", [])
-    fcst_wsd: list[dict] = fcst_info.get("WSD", [])
-    if fcst_vec and fcst_wsd:
-        fcst_wind_text_t1: str = ""
-        fcst_wind_text_t2: str = ""
-        for vec, wsd in zip(fcst_vec, fcst_wsd):
-            fcst_datetime_str: str = vec.get("fcst_datetime_str")
-            # t1/t2 시간 후 예보만 추출
-            if fcst_datetime_str == after_t1_time_str:
-                val_vec_t1: str = vec.get("value", "몰라양")
-                val_vec_t1_text: str = f"{get_wind_direction(val_vec_t1)}"
-                val_wsd_t1: str = wsd.get("value", "몰라양") # 단위: m/s
-                val_wsd_t1_float: float = float(val_wsd_t1)
-                if val_wsd_t1_float >= 4.0 and val_wsd_t1_float < 9.0:
-                    val_wsd_t1_text: str = f"{val_wsd_t1}m/s (약간 강한 바람)"
-                elif val_wsd_t1_float >= 9.0 and val_wsd_t1_float < 14.0:
-                    val_wsd_t1_text: str = f"{val_wsd_t1}m/s (강한 바람)"
-                elif val_wsd_t1_float >= 14.0 and val_wsd_t1_float < 20.0:
-                    val_wsd_t1_text: str = f"{val_wsd_t1}m/s (매우 강한 바람)"
-                elif val_wsd_t1_float >= 20.0:
-                    val_wsd_t1_text: str = f"{val_wsd_t1}m/s (폭풍 수준의 바람)"
-                else:
-                    val_wsd_t1_text: str = f"{val_wsd_t1}m/s"
-                fcst_wind_text_t1 = f"**풍속**: {val_vec_t1_text}풍 {val_wsd_t1_text}\n"
-            elif fcst_datetime_str == after_t2_time_str:
-                val_vec_t2: str = vec.get("value", "몰라양")
-                val_vec_t2_text: str = f"{get_wind_direction(val_vec_t2)}"
-                val_wsd_t2: str = wsd.get("value", "몰라양")
-                val_wsd_t2_float: float = float(val_wsd_t2)
-                if val_wsd_t2_float >= 4.0 and val_wsd_t2_float < 9.0:
-                    val_wsd_t2_text: str = f"{val_wsd_t2}m/s (약간 강한 바람)"
-                elif val_wsd_t2_float >= 9.0 and val_wsd_t2_float < 14.0:
-                    val_wsd_t2_text: str = f"{val_wsd_t2}m/s (강한 바람)"
-                elif val_wsd_t2_float >= 14.0 and val_wsd_t2_float < 20.0:
-                    val_wsd_t2_text: str = f"{val_wsd_t2}m/s (매우 강한 바람)"
-                elif val_wsd_t2_float >= 20.0:
-                    val_wsd_t2_text: str = f"{val_wsd_t2}m/s (폭풍 수준의 바람)"
-                else:
-                    val_wsd_t2_text: str = f"{val_wsd_t2}m/s"
-                fcst_wind_text_t2 += f"**풍속**: {val_vec_t2_text}풍 {val_wsd_t2_text}\n"
-    else:
-        fcst_wind_text_t1: str = ""
-        fcst_wind_text_t2: str = ""
-
-    if fcst_sky_text_t1 == "" and fcst_t1h_text_t1 == "" and fcst_reh_text_t1 == "" and fcst_wind_text_t1 == "":
-        after_text_t1: str = f"--- {time_interval_hour_t1}시간 후 예보 정보가 없어양 ---\n"
-    else:
-        after_head_t1_time: str = (
-            after_t1_time.strftime("%H:%M") 
-            if fcst_base_time.day == after_t1_time.day
-            else after_t1_time.strftime("%m/%d %H:%M")
-        )
-        after_head_t1: str = f"--- {time_interval_hour_t1}시간 후 예보 ({after_head_t1_time}) ---\n"
-        after_text_t1: str = (
-            f"{after_head_t1}"
-            f"{get_fcst_text(fcst_sky_text_t1)}"
-            f"{get_fcst_text(fcst_t1h_text_t1)}"
-            f"{get_fcst_text(fcst_reh_text_t1)}"
-            f"{get_fcst_text(fcst_wind_text_t1)}"
-        )
-    if fcst_sky_text_t2 == "" and fcst_t1h_text_t2 == "" and fcst_reh_text_t2 == "" and fcst_wind_text_t2 == "":
-        after_text_t2: str = f"--- {time_interval_hour_t2}시간 후 예보 정보가 없어양 ---\n"
-    else:
-        after_head_t2_time: str = (
-            after_t2_time.strftime("%H:%M") 
-            if fcst_base_time.day == after_t2_time.day
-            else after_t2_time.strftime("%m/%d %H:%M")
-        )
-        after_head_t2: str = f"--- {time_interval_hour_t2}시간 후 예보 ({after_head_t2_time}) ---\n"
-        after_text_t2: str = (
-            f"{after_head_t2}"
-            f"{get_fcst_text(fcst_sky_text_t2)}"
-            f"{get_fcst_text(fcst_t1h_text_t2)}"
-            f"{get_fcst_text(fcst_reh_text_t2)}"
-            f"{get_fcst_text(fcst_wind_text_t2)}"
-        )
-
-    # embed 메시지 생성
-    embed_title: str = f"{local_address_name}의 날씨 정보에양!"
-    embed_description: str = (
-        f"{ncst_text}\n"
-        f"{after_text_t1}\n"
-        f"{after_text_t2}"
-    )
-    embed_footer: str = (
-        f"위치/날씨 정보 제공: Kakao Local API / 기상청 API (Data.go.kr)\n"
-        f"현재 날씨 시간: {ncst_time}\n"
-        f"예보 발표 시간: {fcst_time}\n"
-        "(날씨 정보는 10분 단위, 예보 정보는 30분 단위로 갱신해양)"
-    )
-
-    embed = discord.Embed(
-        title=embed_title,
-        description=embed_description,
-        color=discord.Colour.from_rgb(135, 206, 235)  # 하늘색
-    )
-    embed.set_footer(text=embed_footer)
-
-    if current_rain_flag:
-        await ctx.send(embed=embed, content=current_rain_text)
-    else:
-        await ctx.send(embed=embed)
-
-@log_command
-async def api_dnf_characters(ctx: commands.Context, server_name: str, character_name: str) -> None:
-    """던전앤파이터 캐릭터 정보 조회
-
-    Args:
-        ctx (commands.Context): Discord 명령어 컨텍스트
-        server_name (str): 서버 이름 (한글)
-        character_name (str): 캐릭터 이름 (특수문자 가능)
-
-    Returns:
-        던전앤파이터 캐릭터 정보 (dict) -> Embed 생성
-
-    Raises:
-        NeopleAPIError: 던전앤파이터 API 요청 중 발생하는 오류
-    """
-    if ctx.message.author.bot:
-        return
-    
-    # 캐릭터 고유 ID 조회
-    try:
-        character_id = neople_dnf_get_character_id(server_name, character_name)
-        server_id = neople_dnf_server_parse(server_name)
-    except NeopleAPIError as e:
-        if "API001" in str(e):
-            await ctx.send(f"네오플 API 요청에 오류가 발생했어양!!!")
-        elif "API002" in str(e):
-            await ctx.send(f"네오플 API 요청 제한에 걸렸어양...")
-        elif "API006" in str(e):
-            await ctx.send(f"네오플 API 요청 파라미터가 잘못되었어양...")
-        elif "DNF000" in str(e):
-            await ctx.send(f"서버명이 잘못 입력 되었어양...")
-        elif "DNF001" in str(e):
-            await ctx.send(f"캐릭터 '{character_name}'을(를) 찾을 수 없어양...")
-        elif "DNF900" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        elif "DNF901" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        elif "DNF980" in str(e):
-            await ctx.send(f"현재 던전앤파이터 서비스 점검 중이에양!")
-        elif "DNF999" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        else:
-            await ctx.send(f"던전앤파이터 API에서 알 수 없는 오류가 발생했어양!")
-        raise NeopleAPIError(str(e))
-
-    # 캐릭터 정보 조회
-    try:
-        request_url = f"{NEOPLE_API_HOME}/df/servers/{server_id}/characters/{character_id}?apikey={NEOPLE_API_KEY}"
-        character_info: dict = general_request_handler_neople(request_url)
-    except NeopleAPIError as e:
-        if "API001" in str(e):
-            await ctx.send(f"네오플 API 요청에 오류가 발생했어양!!!")
-        elif "API002" in str(e):
-            await ctx.send(f"네오플 API 요청 제한에 걸렸어양...")
-        elif "API006" in str(e):
-            await ctx.send(f"네오플 API 요청 파라미터가 잘못되었어양...")
-        elif "DNF000" in str(e):
-            await ctx.send(f"서버명이 잘못 입력 되었어양...")
-        elif "DNF001" in str(e):
-            await ctx.send(f"캐릭터 '{character_name}'을(를) 찾을 수 없어양...")
-        elif "DNF900" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        elif "DNF901" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        elif "DNF980" in str(e):
-            await ctx.send(f"현재 던전앤파이터 서비스 점검 중이에양!")
-        elif "DNF999" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        else:
-            await ctx.send(f"던전앤파이터 API에서 알 수 없는 오류가 발생했어양!")
-        raise NeopleAPIError(str(e))
-
-    # 모험단 이름 추출
-    adventure_name: str = (
-        str(character_info.get("adventureName")).strip()
-        if character_info.get("adventureName") is not None
-        else "adventureNameNotFound"
-    )
-    # 캐릭터 레벨 추출
-    character_level: int = (
-        int(character_info.get("level"))
-        if character_info.get("level") is not None
-        else 0
-    )
-    # 캐릭터 클래스 추출
-    character_job_name: str = (
-        str(character_info.get("jobName")).strip()
-        if character_info.get("jobName") is not None
-        else "몰라양"
-    )
-    # 캐릭터 전직명 추출
-    character_job_grow_name: str = (
-        str(character_info.get("jobGrowName")).strip()
-        if character_info.get("jobGrowName") is not None
-        else "몰라양"
-    )
-    # 캐릭터 명성 추출
-    character_fame: int = (
-        int(character_info.get("fame"))
-        if character_info.get("fame") is not None
-        else 0
-    )
-    # 캐릭터 길드 추출
-    character_guild: str = (
-        str(character_info.get("guildName")).strip()
-        if character_info.get("guildName") is not None
-        else "길드가 없어양!"
-    )
-
-    dundam_url = f"https://dundam.xyz/character?server={server_id}&key={character_id}"
-    dfgear_url_c = f"https://dfgear.xyz/character?sId={server_id}&cId={character_id}&cName={character_name}"
-    if adventure_name != "adventureNameNotFound":
-        dfgear_url_a = f"https://dfgear.xyz/adventure?cName={adventure_name}"
-        dfgear_url_desc = (
-            f"[🔗 DFGEAR 사이트 이동 (캐릭터)]({dfgear_url_c})\n"
-            f"[🔗 DFGEAR 사이트 이동 (모험단)]({dfgear_url_a})\n"
-        )
-    else:
-        dfgear_url_desc = f"[🔗 DFGEAR 사이트 이동]({dfgear_url_c})\n"
-
-    embed_description: str = (
-        f"[🔗 던담 사이트 이동]({dundam_url})\n"
-        f"{dfgear_url_desc}"
-        f"**모험단:** {adventure_name}\n"
-        f"**레벨:** {character_level}\n"
-        f"**직업:** {character_job_name}\n"
-        f"**전직:** {character_job_grow_name}\n"
-        f"**명성:** {character_fame}\n"
-        f"**길드:** {character_guild}\n"
-    )
-    embed_footer: str = (
-        f"캐릭터 선택창에 나갔다 오면 빨리 갱신되양!\n"
-        f"powered by Neople API"
-    )
-
-    # 캐릭터 이미지 URL추출
-    character_image_url = f"https://img-api.neople.co.kr/df/servers/{server_id}/characters/{character_id}?zoom=1"
-    character_image_bytes: io.BytesIO = get_image_bytes(character_image_url)
-    today_date_str: str = datetime.now().strftime("%Y%m%d%H%M")
-    character_image_filename = f"{server_id}_{character_id}_{today_date_str}.png"
-    buffer = discord.File(character_image_bytes, filename=character_image_filename)
-    # Discord Embed 객체 생성
-    if character_job_name == "마법사(여)":
-        embed_color = discord.Colour.from_rgb(255, 0, 0)  # red
-    else:
-        embed_color = discord.Colour.from_rgb(128, 128, 128)  # grey
-    embed = discord.Embed(
-        title=f"{server_name}서버 '{character_name}' 모험가님의 정보에양!",
-        description=embed_description
-    )
-    embed.set_footer(text=embed_footer)
-    embed.colour = embed_color
-    embed.set_image(url=f"attachment://{character_image_filename}")
-
-    # Discord Embed 전송
-    await ctx.send(embed=embed, file=buffer)
-    buffer.close()
-
-@log_command
-async def api_dnf_timeline_weekly(ctx: commands.Context, server_name: str, character_name: str) -> None:
-    """던전앤파이터 캐릭터 주간 타임라인 조회 (이번주 기준)
-
-    Args:
-        ctx (commands.Context): Discord context
-        server_name (str): 서버 이름
-        character_name (str): 캐릭터 이름
-
-    Raises:
-        NexonAPIBadRequest: 잘못된 요청
-        NexonAPIForbidden: 접근 금지
-        Exception: API 요청 오류
-        Exception: API 응답 오류
-        Exception: 데이터 처리 오류
-        Exception: 기타 오류
-        Exception: 알 수 없는 오류
-
-    Note:
-        타임라인 기간: 이번주 목요일 오전 6시 ~ 현재시간 (최대 차주 목요일까지)
-    """
-    try:
-        timeline_data: dict = get_dnf_weekly_timeline(server_name, character_name)
-    except NeopleAPIError as e:
-        if "API001" in str(e):
-            await ctx.send(f"네오플 API 요청에 오류가 발생했어양!!!")
-        elif "API002" in str(e):
-            await ctx.send(f"네오플 API 요청 제한에 걸렸어양...")
-        elif "API006" in str(e):
-            await ctx.send(f"네오플 API 요청 파라미터가 잘못되었어양...")
-        elif "DNF000" in str(e):
-            await ctx.send(f"서버명이 잘못 입력 되었어양...")
-        elif "DNF001" in str(e):
-            await ctx.send(f"캐릭터 '{character_name}'을(를) 찾을 수 없어양...")
-        elif "DNF900" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        elif "DNF901" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        elif "DNF980" in str(e):
-            await ctx.send(f"현재 던전앤파이터 서비스 점검 중이에양!")
-        elif "DNF999" in str(e):
-            await ctx.send(f"던전앤파이터 API에서 오류가 발생했어양!")
-        else:
-            await ctx.send(f"던전앤파이터 API에서 알 수 없는 오류가 발생했어양!")
-        raise NeopleAPIError(str(e))
-    except NeopleDNFInvalidTimelineParams as e:
-        await ctx.send(f"타임라인을 불러오는데 문제가 발생했어양!")
-        raise Exception(str(e))
-    except NeopleDNFInvalidCharacterInfo as e:
-        await ctx.send(f"캐릭터 '{character_name}'을(를) 찾을 수 없어양...")
-
-    character_timeline: dict = timeline_data.get("timeline")
-    timeline_rows: List[Dict[str, Any]] = character_timeline.get("rows")
-    if len(timeline_rows) == 0:
-        await ctx.send(f"이번주에 레전더리 이상 등급의 득템 기록이나, 레이드/레기온 클리어 기록이 없어양!")
-        return
-    
-    else:
-        # timeline 시간 내림차순으로 데이터가 정렬되어 있음
-
-        # 캐릭터 기본 정보 추출
-        adventure_name: str = timeline_data.get("adventureName", "몰라양")
-        level: int = timeline_data.get("level", 0)
-        job_name: str = timeline_data.get("jobName", "몰라양")
-        job_grow_name: str = timeline_data.get("jobGrowName", "몰라양")
-        fame: int = timeline_data.get("fame", 0)
-
-        # timeline 데이터 생성
-        timeline_title: str = f"{server_name}서버 '{character_name}' 모험가님의 이번주 주간던파에양!"
-        timeline_highlight: str = ""
-        get_legendary_count: int = 0
-        get_epic_count: int = 0
-        get_epic_up_count: int = 0 # 융합석 장비 업그레이드 횟수
-        get_primeval_count: int = 0
-        clear_raid_twilight_flag: bool = False
-        clear_raid_nabel_flag: bool = False
-        clear_raid_mu_flag: bool = False
-        clear_raid_region_flag: bool = False
-
-        # 타임라인 데이터 파싱
-        for row in timeline_rows:
-            timeline_code: int = row.get("code")
-            timeline_name: str = row.get("name")
-            timeline_date: str = row.get("date") #YYYY-MM-DD HH:MM
-            timeline_data: dict[str, Any] = row.get("data")
-
-            # 아이템 획득
-            if 600 > timeline_code >= 500:
-                item_name: str = timeline_data.get("itemName", "몰라양")
-                item_rare: str = timeline_data.get("itemRarity", "몰라양")
-
-                # 태초 아이템 획득 시 하이라이트 메시지 생성
-                if timeline_code != 513 and item_rare == "태초":
-                    channel_name = timeline_data.get("channelName", "알수없음")
-                    channel_no = timeline_data.get("channelNo", "알수없음")
-                    get_primeval_count += 1
-                    timeline_highlight += (
-                        f"{channel_name} {channel_no}채널에서 {dnf_convert_grade_text(item_rare)}{item_name} 아이템을 획득했어양! ({timeline_date})\n"
-                    )
-
-                if timeline_code == 513 and item_rare == "태초":
-                    # 던전 카드 보상에서 태초 아이템 획득 시
-                    dungeon_name: str = timeline_data.get("dungeonName", "몰라양")
-                    get_primeval_count += 1
-                    timeline_highlight += (
-                        f"던전 {dungeon_name}에서 카드 보상으로 {dnf_convert_grade_text(item_rare)}{item_name} 아이템을 획득했어양! ({timeline_date})\n"
-                    )
-
-                # 융합석 업그레이드 획득 시 (에픽 획득 집계 미포함)
-                if timeline_code == 511 and item_rare == "에픽":
-                    get_epic_up_count += 1
-                    timeline_highlight += (
-                        f"융합석 업글레이드를 통해 {dnf_convert_grade_text(item_rare)}{item_name} 아이템을 획득했어양! ({timeline_date})\n"
-                    )
-                
-                # 에픽 아이템 획득
-                if item_rare == "에픽":
-                    get_epic_count += 1
-
-                # 레전더리 아이템 획득
-                if item_rare == "레전더리":
-                    get_legendary_count += 1
-
-            if timeline_code == 209:
-                # 레기온 클리어
-                region_name: str = timeline_data.get("regionName", "몰라양")
-                if region_name == "베누스":
-                    clear_raid_region_flag = True
-                    clear_raid_region_date = timeline_date
-
-            if timeline_code == 201:
-                # 레이드 클리어
-                raid_name: str = timeline_data.get("raidName", "몰라양")
-                if raid_name == "이내 황혼전":
-                    clear_raid_twilight_flag = True
-                    clear_raid_twilight_date = timeline_date
-                if raid_name == "만들어진 신 나벨":
-                    clear_raid_nabel_flag = True
-                    clear_raid_nabel_date = timeline_date
-                if raid_name == "아스라한":
-                    clear_raid_mu_flag = True
-                    clear_raid_mu_date = timeline_date
-
-            # 아이템 증폭
-            if timeline_code == 402:
-                if "증폭" in timeline_name:
-                    up_type = "증폭"
-                elif "강화" in timeline_name:
-                    up_type = "강화"
-                elif "제련" in timeline_name:
-                    up_type = "제련"
-                else:
-                    raise Exception("Invalid upgrade type in timeline data")
-                
-                up_item_rare: str = timeline_data.get("itemRarity", "몰라양")
-                up_item_name: str = timeline_data.get("itemName", "몰라양")
-                up_item_before: int = timeline_data.get("before", 0)
-                up_item_after: int = timeline_data.get("after", 0)
-                up_item_result: bool = timeline_data.get("result", False)
-                up_item_safe: bool = timeline_data.get("safe", False)
-
-                # 보호권 사용 여부 텍스트
-                if up_item_safe:
-                    up_safe_text: str = "증폭/강화 보호권 사용"
-                else:
-                    up_safe_text: str = "증폭/강화 보호권 미사용"
-
-                if up_item_before >= 10:
-                    # 10강 이상 증폭/강화 시 하이라이트 메시지 생성
-                    timeline_highlight += (
-                        f"{dnf_convert_grade_text(up_item_rare)} {up_item_name} {up_item_after} {up_type}에 "
-                        f"{'성공' if up_item_result else '실패'} 했어양! ({timeline_date})\n"
-                    )
-                
-                if up_item_after == 8 and up_type =="제련" and up_item_result:
-                    # 8제련 성공 시 하이라이트 메시지 생성
-                    timeline_highlight += (
-                        f"{dnf_convert_grade_text(up_item_rare)} {up_item_name} 8 제련에 "
-                        f"성공 했어양! ({timeline_date})\n"
-                    )
-
-        # 타임라인 요약 메시지 생성
-        if timeline_highlight != "":
-            timeline_highlight_str: str = f"**\-\-\- 주간 하이라이트 \-\-\-**\n{timeline_highlight}\n"
-        else:
-            timeline_highlight_str: str = ""
-
-        clear_raid_twilight = dnf_get_clear_flag(clear_raid_twilight_flag, locals().get('clear_raid_twilight_date'))
-        clear_raid_nabel = dnf_get_clear_flag(clear_raid_nabel_flag, locals().get('clear_raid_nabel_date'))
-        clear_raid_mu = dnf_get_clear_flag(clear_raid_mu_flag, locals().get('clear_raid_mu_date'))
-        clear_raid_region = dnf_get_clear_flag(clear_raid_region_flag, locals().get('clear_raid_region_date'))
-
-        timeline_summary: str = (
-            f"모험단명: {adventure_name}\n"
-            f"레벨: {level}\n"
-            f"직업: {job_name}, {job_grow_name}\n"
-            f"명성: {fame:,}\n\n"
-            f"**\-\-\- 이번주 장비 획득 \-\-\-**\n"
-            f"🟢 태초 획득: {get_primeval_count}개\n"
-            f"🟡 에픽 획득: {get_epic_count}개 (융합석 업글 {get_epic_up_count}회)\n"
-            f"🟠 레전 획득: {get_legendary_count}개\n\n"
-            f"**\-\-\- 레이드 및 레기온 클리어 현황 \-\-\-**\n"
-            f"이내 황혼전 클리어: {clear_raid_twilight}\n"
-            f"만들어진 신 나벨 클리어: {clear_raid_nabel}\n"
-            f"아스라한 클리어: {clear_raid_mu}\n"
-            f"베누스 레기온 클리어: {clear_raid_region}\n"
-            f"\n{timeline_highlight_str}"
-        )
-
-        timeline_footer: str = (
-            f"목요일 오전 6시 이후 집계\n"
-            f"융합석 업그레이드는 에픽 획득에 포함되지 않아양\n"
-            f"powered by Neople API"
-        )
-
-        # Discord Embed 객체 생성
-        embed = discord.Embed(
-            title=timeline_title,
-            description=timeline_summary
-        )
-        embed.set_footer(text=timeline_footer)
-        embed.colour = discord.Colour.from_rgb(128, 0, 128)  # purple
-        await ctx.send(embed=embed)
-
 
 @log_command
 async def api_maple_fortune_today(ctx: commands.Context, character_name: str) -> None:
@@ -1760,6 +1027,7 @@ async def api_maple_fortune_today(ctx: commands.Context, character_name: str) ->
     )
     embed.set_footer(text=embed_footer)
     await ctx.send(embed=embed)
+
 
 @log_command
 async def api_maple_xp_history(ctx: commands.Context, character_name: str) -> None:
