+++ service/yfinance_command.py
@@ -6,50 +6,18 @@
 
 """
 import discord
-from yfinance import Ticker
-import requests
-from bs4 import BeautifulSoup
 from discord.ext import commands
-from service.common import log_command
-from pytz import timezone
+
+from yfinance import Ticker
+
+from service.yfinance_utils import exchange_krw_rate
 from datetime import datetime
+from pytz import timezone
 
-from service.common import safe_float, safe_percent
-from service.stk_exception import *
+from bot_logger import log_command
+from utils.text import safe_float, safe_percent
 
-def exchange_krw_rate(from_currency: str) -> float:
-    """
-    환율을 검색해서 (KRW)로 변환하는데 도움을 줍니다.
-
-    Args:
-        from_currency (str): 변환할 통화의 코드 (예: 'USD', 'EUR')
-        amount (float): 변환할 금액
-
-    Returns:
-        float: 변환된 금액 (원화)
-
-    Raises:
-        Exception: 환율 정보를 가져오는 중 오류가 발생한 경우
-    """
-    from_currency = from_currency.strip().upper()
-    if from_currency == 'KRW':
-        return 1
-    else:
-        url = f"https://finance.naver.com/marketindex/exchangeDetail.naver?marketindexCd=FX_{from_currency}KRW"
-        response = requests.get(url)
-        html = BeautifulSoup(response.text, 'html.parser')
-
-    if response.status_code == 200:
-        # 환율 정보를 가져오는 데 성공한 경우
-        options = html.select("select.selectbox-source option")
-        for opt in options:
-            if from_currency in opt.text:
-                rate: float = float(opt.get("value"))
-                return rate
-        raise STK_ERROR_NO_RATE(f"환율 정보를 찾을 수 없어양: {from_currency}")
-    else:
-        # 환율 정보를 가져오는 데 실패한 경우
-        raise STK_ERROR_FETCH_RATE(f"HTTP {response.status_code}: {response.reason}")
+from exceptions.api_exceptions import *
 
 @log_command
 async def stk_us_stock_price(ctx: commands.Context, ticker: str) -> float:
@@ -66,7 +34,7 @@ async def stk_us_stock_price(ctx: commands.Context, ticker: str) -> float:
         previous_close = stock.info.get('regularMarketPreviousClose', '몰라양')
         today_close = stock.info.get('regularMarketPrice', '몰라양')
         if previous_close == '몰라양' or today_close == '몰라양':
-            raise STK_ERROR_NO_TICKER(f'티커 {ticker}는 존재하지 않거나, 주식 정보가 없어양!')
+            raise YFI_NO_TICKER(f'티커 {ticker}는 존재하지 않거나, 주식 정보가 없어양!')
         stock_name = stock.info.get('shortName', ticker)
         stock_ticker = stock.info.get('symbol', ticker)
         stock_timezone = stock.info.get('exchangeTimezoneName', 'America/New_York')
@@ -102,12 +70,12 @@ async def stk_us_stock_price(ctx: commands.Context, ticker: str) -> float:
             low_52w_krw_text = ""
             footer_text_extra = ""
 
-    except STK_ERROR_NO_TICKER:
+    except YFI_NO_TICKER:
         await ctx.send(f"티커 {ticker}는 존재하지 않거나, 주식 정보가 없어양!")
         return
-    except STK_ERROR_NO_RATE:
+    except YFI_NO_RATE_WARNING:
         await ctx.send(f"경고) {stock_currency} 환율 정보를 찾을 수 없어양!")
-    except STK_ERROR_FETCH_RATE:
+    except YFI_STOCK_FETCH_RATE:
         await ctx.send(f"경고) 환율 정보를 가져오는 데 실패했어양!")
     
     finally:
